% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{FPCA}
\alias{FPCA}
\title{Functional principal component analysis}
\usage{
FPCA(
  basis_mat,
  knot_space,
  data_projected_name = paste0("data_projected_", knot_space, "mm.dat"),
  train_sub,
  test_sub
)
}
\arguments{
\item{basis_mat}{matrix of basis functions (each basis is vectorised in one column of the matrix)}

\item{knot_space}{space between consecutive knots (in mm, equal for each dimension)}

\item{data_projected_name}{text file with the smoothing projections for each statistical unit}

\item{train_sub}{rownames of \code{data_projected_name} used to fit the FPCA}

\item{test_sub}{rownames of \code{data_projected_name} for which to get the FPCA scores}

\item{pve_threshold}{proportion of variance explained}
}
\value{
A list with the following elements
\describe{
  \item{scores}{Scores matrix (each row is for one statistical unit)}
  \item{eFuns}{Eigenimages (vectorised on each column)}
  \item{eVals}{Eigenvalues}
}
}
\description{
Return eigenimages, eigenvectors and scores.
}
\author{
Marco Palma, \email{M.Palma@warwick.ac.uk}
}
\keyword{FPCA}
