% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Colour palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colours}
}
\description{
Create colour palette with equidistant hues.
}
\examples{
gg_color_hue(3)

}
\author{
Marco Palma, \email{M.Palma@warwick.ac.uk}
}
\keyword{gg_color_hue}
