% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predint_plot.R
\name{predint_plot}
\alias{predint_plot}
\title{Illustration of crayon colors}
\usage{
predint_plot(pred_interval, mycol = c("#009E73", "gold3", "#D55E00"),
  xlab_input = "Difference from chronological age",
  ylab_input = "Subjects")
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{hplot}
