% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_image.R
\name{resize_image}
\alias{resize_image}
\title{Select active areas within a mask}
\usage{
resize_image(mask, img = mask)
}
\arguments{
\item{mask}{path of the mask file}

\item{img}{path of the image file}
}
\value{
A list with the following elements
\describe{
  \item{array}{array containing voxels within the mask}
  \item{original_coord}{coordinates of the limits of the mask with respect to the original size of the image}
}
}
\description{
Select the part of the image within the mask.
}
\author{
Marco Palma, \email{M.Palma@warwick.ac.uk}
}
\keyword{resize_image}
