% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_projection_Bspl.R
\name{calc_projection_Bspl}
\alias{calc_projection_Bspl}
\title{Bspline projection of 3D images}
\usage{
calc_projection_Bspl(knot_space, mask_fname)
}
\arguments{
\item{knot_space}{space between consecutive knots (in mm, equal for each dimension)}

\item{mask_fname}{path of the mask file}
}
\value{
A list with the following elements
\describe{
  \item{basis_mat}{Matrix of basis functions (each basis is vectorised in one column of the matrix)}
  \item{voxel_grid_nonzero_mask}{indices of voxels within the mask}
  \item{dims_mask}{dimensions of the mask}
  \item{mask_subset}{resized mask}
}
}
\description{
Create a design matrix from an isotropic tensor product of Bsplines only 
for voxels within a mask.
}
\author{
Marco Palma, \email{M.Palma@warwick.ac.uk}
}
\keyword{calc_projection_Bspl}
