% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqr_prediction.R
\name{fqr_prediction}
\alias{fqr_prediction}
\title{Illustration of crayon colors}
\usage{
fqr_prediction(train_sub, test_sub, data_demo, pred_table,
  qr_pen = "LASSO", qr_postLASSO = FALSE, lambda = NULL, tau_levels,
  return_func_coef = FALSE)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{hplot}
