% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predint_plot.R
\name{predint_plot}
\alias{predint_plot}
\title{Prediction intervals for functional quantile regression}
\usage{
predint_plot(
  pred_interval,
  col = c("#009E73", "gold3", "#D55E00"),
  xlab_input = "Difference from chronological age",
  ylab_input = "Subjects"
)
}
\arguments{
\item{pred_interval}{rownames of \code{data_projected_name} used to fit the FPCA}

\item{col}{rownames of \code{data_projected_name} for which to get the FPCA scores}

\item{xlab_input}{text file with the smoothing projections for each statistical unit}

\item{ylab_input}{demographic data with the scalar outcome of interest}
}
\description{
Return point predictions and prediction intervals for functional 
scalar-on-image quantile regression and outputs from FPCA.
}
\author{
Marco Palma, \email{M.Palma@warwick.ac.uk}
}
\keyword{FPCA}
