% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slices_plot.R
\name{slices_plot}
\alias{slices_plot}
\title{Select active areas within a mask}
\usage{
slices_plot(
  image_vec,
  mask_obj = mask,
  dims = dims_mask,
  img_templ = img_template,
  voxels,
  col_threshold = 0,
  legend_range = range(image_vec),
  slice_number = FALSE,
  ...
)
}
\arguments{
\item{image_vec}{vectorised image values}

\item{mask_obj}{original mask image}

\item{dims}{image dimensions}

\item{img_templ}{minimal deformation template (for TBM images)}

\item{voxels}{voxels indices within the mask (it must have the same dimensions as image_vec)}

\item{col_threshold}{midpoint in the divergent colour scale}

\item{legend_range}{extremes of the divergent colour scale}

\item{slice_number}{add a note about the slices number in the array}
}
\description{
Select the part of the image within the mask.
}
\author{
Marco Palma, \email{M.Palma@warwick.ac.uk}
}
\keyword{slices_plot}
