% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slices_plot.R
\name{slices_plot}
\alias{slices_plot}
\title{Illustration of crayon colors}
\usage{
slices_plot(image_vec, dims = dims_mask,
  voxels = results$voxel_grid_nonzero_mask, col_threshold = 0,
  legend_range = range(image_vec), slice_number = FALSE, ...)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()

}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{hplot}
